fix(retry decorator): allow retry if retries = 1
if we use @retry(retries=1) we assume that we will try to run function and retry one more time.
But range(1) will generate just single element, and actually we will not have chance to retry.

By default, we had "try" + "2 retries". To keep it we should decrease default retries value.